+++ lib/surface/directive/events.ex
@@ -9,12 +9,15 @@ defmodule Surface.Directive.Events do
     "focus",
     "change",
     "submit",
+    "auto-recover",
     "keydown",
     "keyup",
     "window-focus",
     "window-blur",
     "window-keydown",
-    "window-keyup"
+    "window-keyup",
+    "viewport-top",
+    "viewport-bottom"
   ]
 
   @phx_events Enum.map(@events, &"phx-#{&1}")