+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 This version moves all Form helpers to a [separate library](https://github.com/surface-ui/surface_form). Form Helpers are no longer used in new apps from Surface v0.12. Older applications who wish to maintain compatibility, add `{:surface_form, "~> 0.1.0"}` to your `mix.exs`.
 
+* Add support for the `:on-auto-recover`, `:on-viewport-top` and `:on-viewport-bottom` event directives
+* Move form-specific event logic (`phx-trigger-action`) to `surface_form`
+
 ## v0.11.5 (2024-09-04)
 
 * Support Liveview `~> 0.20.15` (#749)